More work on 2D slice of 3D world
The problem: We can't simply take a 2D slice out of a 3D world because objects are only rendered by their outer rim. For example, if a sphere passes through our 2D world, only the outer rim of the colliding part is going to be visible (a thin empty round layer will be seen). This happens because only the outer triangles creating the sphere really exist, the inside of the sphere doesn't exist.

Generally we came up with 3 solutions for taking a 2D slice out of a 3D world.
Assuming we have a 2D plane colliding with a 3D scene:
1. Using code to find the vertices of collision between the plane and any 3D object and filling the plane with triangles accordingly.
Generally a good solution for the problem, but VERY performance heavy.
2. Implementing a shader to render a 2D version of every 3D object to our needs in real time. Hard to implement, some performance issues.
3. Making an illusion of a 2D world by placing a 2D plane visible only by the 2D camera, making object coming from behind "suddenly" appearing into the 2D scene as if from a different dimension.
The problem is objects coming from the front has to be of specific shape and color for the illusion not to break.